# drivers/neopixel/neopixel.py: Optimize fill() and reduce code size.