@robert-hh pointed out the subtle behavior of the fill method here: https://github.com/micropython/micropython/commit/71f4faac2732d932dcc03bdbc9f80434e5757edb#commitcomment-55180052

Although it works, this can be optimized significantly.

For a 200-LED RGB strip, a single fill() was taking 14.4ms on PYBV11 (12.4ms on SF6). It now takes 2.0ms (and 1.5ms). For comparison a write takes 1.1ms. This costs 40 bytes (bytecode).

The second commit does some renaming to reduce generated code size by 83 bytes, so the net impact of this PR is negative.